requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0

# TUI dependencies
textual>=6.6.0
